code refactoring, improvements and fixes
