+++ assets/js/lib/compileMarkdown.js
@@ -1,0 +1,4 @@
+import marked from 'marked'
+import DOMPurify from 'dompurify'
+
+export default (html) => html ? DOMPurify.sanitize(marked(html)) : ""
